+++ server/server.ts
@@ -25,6 +25,7 @@ import noteViewRouter from './routers/note-view.js'
 import dashboardRouter from './routers/dashboard.js'
 import serachProfileRouter from './routers/search-profile.js'
 import settingsRouter from './routers/settings.js'
+import apiRouter  from './services/apis.js';
 
 import Notes from './schemas/notes.js'
 import Students from './schemas/students.js'
@@ -81,6 +82,7 @@ app.use('/view', noteViewRouter(io))
 app.use('/dashboard', dashboardRouter(io))
 app.use('/search-profile', serachProfileRouter(io))
 app.use('/settings', settingsRouter(io))
+app.use('/api', apiRouter(io))
 app.use(errorHandler) // Middleware for handling errors
 
 app.get('/logout', (req, res) => {
@@ -111,103 +113,6 @@ app.get('/signup2', (req, res) => {
 })
 app.get('/login2', (req, res) => {
     res.render('login2')
-})
-
-app.post('/download', async (req, res) => {
-    let noteID = req.body.noteID
-    let noteTitle = req.body.noteTitle
-
-    const noteLinks = (await Notes.findById(noteID, { content: 1 })).content
-
-    res.setHeader('Content-Type', 'application/zip');
-
-    const sanitizeFilename = (filename) => {
-        return filename.replace(/[^a-zA-Z0-9\.\-\_]/g, '_'); // Replace any invalid characters with an underscore
-    }
-    res.setHeader('Content-Disposition', `attachment; filename=${sanitizeFilename(noteTitle)}.zip`);
-
-    const archive = archiver('zip', { zlib: { level: 9 } });
-    archive.pipe(res);
-
-    for (let imageUrl of noteLinks) {
-        try {
-            const response = await fetch(imageUrl);
-            if (!response.ok) {
-                throw new Error(`Failed to fetch ${imageUrl}: ${response.statusText}`);
-            }
-
-            let arrayBuffer = await response.arrayBuffer()
-            let buffer = Buffer.from(arrayBuffer)
-
-            let fileName = basename(new URL(imageUrl).pathname);
-            archive.append(buffer, { name: fileName });
-        } catch (err) {
-            console.error(`Error fetching image: ${err.message}`);
-            return;
-        }
-    }
-    archive.finalize();
-})
-
-app.get('/search', async (req, res, next) => {
-    let searchTerm = req.query.q as string
-    const regex = new RegExp(searchTerm.split(' ').map(word => `(${word})`).join('.*'), 'i');
-    let notes = await Notes.find({ title: { $regex: regex } }, { title: 1 })
-    res.json(notes)
-})
-
-app.get('/searchUser', async (req, res) => {
-    if (req.query) {
-        let term = req.query.term
-        let students = await Students.aggregate([
-            {
-                $match: {
-                    displayname: { $regex: `^${term}`, $options: 'i' } // Ensure case-insensitive search
-                }
-            },
-            {
-                $project: {
-                    displayname: 1,
-                    username: 1,
-                    profile_pic: 1,
-                    _id: 0
-                }
-            }
-        ]);
-
-        res.json(students)
-    }
-})
-
-app.get('/getnote', async (req, res, next) => {
-    let type = req.query.type
-    async function getSavedNotes(studentDocID) {
-        let student = await Students.findById(studentDocID, { saved_notes: 1 })
-        let saved_notes_ids = student['saved_notes']
-        let notes = await Notes.find({ _id: { $in: saved_notes_ids } }, { title: 1 })
-        return notes
-    }
-    if (type === 'save') {
-        let studentDocID = req.query.studentDocID
-        let savedNotes = await getSavedNotes(studentDocID)
-        res.json(savedNotes)
-    } else if (type === 'seg') {
-        let page = req.query.page as unknown as number
-        let count = req.query.count as unknown as number
-        let skip: number = (page - 1) * count
-        let notes = await Notes.find({}).sort({ createdAt: -1 }).skip(skip).limit(count).populate('ownerDocID', 'profile_pic displayname studentID')
-        if (notes.length != 0) {
-            res.json(notes)
-        } else {
-            res.json([])
-        }
-    }
-})
-
-app.get('/getNotifs', async (req, res) => {
-    let studentID = req.query.studentID?.toString()
-    let notifs = await getNotifications(studentID)
-    res.json(notifs)
 })
 
 app.get('/message', async (req, res) => {
